+++ presentation/src/main/res/layout/fragment_memo_list.xml
@@ -8,7 +8,7 @@
         xmlns:tools="http://schemas.android.com/tools"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".memo.MemoListFragment">
+        tools:context=".memo.fragment.MemoListFragment">
 
         <com.google.android.material.progressindicator.CircularProgressIndicator
             android:id="@+id/progressIndicator_memoList"
